Fix reported source of a test
What matters is where the `it` and the test function are called and
defined.  If I used `nio.tests.it` they are defined inside the nio
library, so the source of the test will be reported there.

We still need an asynchronous context when calling asynchronous
functions.  I use a wrapper function for the asynchronous part.  The
test is defined in the proper location using `it` and the wrapper
ensures that the asynchronous code can be run inside an asynchronous
context.