feat: handling multilines results with custom output handler
Lazily `require` JSON module in output handler

We `require` the output handler in Neovim in order to get the full path
to its source.  This executes all top-level statements of the handler,
including any `require` calls.  If one of the required modules is not
installed (in this case `dkjson`) an error is raised.

The solution is to move the `require` call into the function call of the
module.  The module will not be called by Neovim, but by busted, and the
dkjson module will be present in that context.

Add commentary (motivation) to output handler

Undo stylistic changes

Shorten name of `output_handler_path` variable

Makes it easier to read.

Search for last result line containing the marker

Rename marker variable

The variable is already scoped to the output handler, so the name can be
something short.

Add tests which write to standard output

These tests are meant to be run from within Neovim.

Add test for when output contains result marker

This is why it is important to search for the output marker backwards.